refactor to make it easier to add new kinds of UDP handlers. add UDP handler for CCT bulbs
